refactor: rename src/scss folder to src/build
