+++ index.js
@@ -1,9 +1,9 @@
 const embedFile = require('./embedFile');
-const sass = require('./sass');
+const build = require('./build');
 const utils = require('./utils');
 const components = require('./components');
 
 module.exports.embedFile = embedFile;
-module.exports.sass = sass;
+module.exports.build = build;
 module.exports.utils = utils;
 module.exports.components = components;
